admin/requirements.txt
sanic
mysql-connector-python
requests